==> code/07_rando_checks.ipynb <==
import pandas as pd

#Here is a list of indoor monitors, for your edification
site_info = pd.read_csv('../GIS/purpleair/houmetro-pa-sensors-atleast_1yr.csv')
indoor_sensor_list = site_info[site_info['location_type'] == 1]['sensor_index'].tolist()
indoor_sensor_list

#testing tx-wide EPA data
tx_epa_23 = pd.read_csv('../data/source/EPA/2023_tx_epa_ad_viz_plotval_data.csv')

epa_23_by_county = tx_epa_23.groupby('County',dropna=False).agg(min_date=('Date','min'),
                                                                max_date=('Date','max'),
                                                                avg_epa_reading=('Daily Mean PM2.5 Concentration','mean')
                                                                ).reset_index()

epa_23_by_county.to_csv('../data/analyzed/2023_tx_epa_by_county.csv',index=False)
